Move UtilBitScanForward to Util.h
